refactor(agent): share text-scanning helpers across NL request parsers
The natural-language request parsers (the mock/simulated tool-call path)
each copy-pasted the same private text-scanning helpers. Extracted the
verified byte-identical ones into a shared AgentRequestTextScanner:

- backtickQuotedValues(in:) — was byte-identical in 4 files
  (AgentFileReadRequestParser, AgentDownloadRequestParser,
  AgentFileListRequestParser, MockLLMClient).
- alphanumericWordTokens(in:) — the Set<String> word tokenizer, byte-
  identical in AgentGitReadRequestParser + AgentWorkspaceDiagnosticRequestParser.

Deliberately NOT merged (different behavior, left in place): FileList's
own tokenizeWords keeps apostrophes in words, and MockPullRequestIntentPlanner's
returns an ordered [String] — both are genuinely different rules, so the
shared helper documents that parsers needing them keep their own.

Behavior-preserving; every call site repointed; full suite green.
